Commit after part D

+++ my_shell.c
@@ -4,17 +4,15 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <sys/stat.h>
 
 #define MAX_INPUT_SIZE 1024
 #define MAX_TOKEN_SIZE 64
 #define MAX_NUM_TOKENS 64
 
 int noOfTokens = 0; 
-//pid_t childPid = 0;
-//pid_t bgChildPid = 0;
-pid_t fgPid;
+pid_t fgPid = -1;
 pid_t bgPid;
-//pid_t grandChildPid = 0;
 int processes[1000];
 int noOfProcesses = 0;
 
@@ -59,12 +57,29 @@ char **tokenize(char *line)
 	return tokens;
 }
 
+int isValidDirectory(char *dirName)
+{
+	struct stat stats;
+    stat(dirName, &stats);
+    if (S_ISDIR(stats.st_mode))
+	{
+		return 1;
+	}  
+    return 0;
+}
+
 void exitFunction(char **tokens)
 {
 	for (int i = 0; i < noOfProcesses; i++)
 	{
-		//printf("\n Process %d %d", i, processes[i]);
-		kill(processes[i], SIGKILL);
+		if (processes[i] > 0)
+		{
+			//printf("\n Process %d %d", i, processes[i]);
+			kill(processes[i], SIGKILL);
+			waitpid(processes[i], NULL, 0);
+			//printf("PID : %d is now reaped.\n", processes[i]);
+			processes[i] = -10;
+		}	
 	}
 
 	for(int i=0;tokens[i]!=NULL;i++)
@@ -82,12 +97,13 @@ void executeCommand(char **tokens)
 {
 	int status=0;
 	pid_t childPid = fork();
-    fgPid = childPid;
+    //fgPid = childPid;
+
 	//child process
 	if (childPid == 0)
 	{
 		//to check for child process in ps -eF | grep './my_shell'
-		sleep(20);
+		//sleep(10);
 		
 		/*
 		To check when the backgroud process is returning faster than foreground process, 
@@ -101,18 +117,35 @@ void executeCommand(char **tokens)
 	}
 	else
 	{
+		fgPid = childPid;
         waitpid(childPid, &status, WUNTRACED);
 	}
 }
 
 void childSignalHandler(int signum) {
 	int status;
-	pid_t return_pid = waitpid(bgPid, &status, WNOHANG);
+	// pid_t return_pid = waitpid(bgPid, &status, WNOHANG);
 
-	if (return_pid == bgPid) 
-	{
-		printf("Shell: Background process finished\n");
-	}
+	// if (return_pid == bgPid) 
+	// {
+	// 	printf("Shell: Background process finished\n");
+		for (int i = 0; i < noOfProcesses; i++)
+		{
+			if(processes[i] > 0)
+			{
+				pid_t return_pid = waitpid(processes[i], &status, WNOHANG);
+				if (return_pid == processes[i]) 
+				{
+					printf("Shell: Background process finished (PID : %d)\n",processes[i]);
+					if (processes[i] == bgPid)
+					{
+						processes[i] = -10;
+					}
+				}
+			}
+		}
+		
+	// }
 
 }
 
@@ -120,19 +153,20 @@ void executeCommandInBackground(char **tokens)
 {
 	int status;
 	pid_t bgChildPid = fork();
-    bgPid = bgChildPid;
+    
 	setpgid(0,0);
 	//printf("Background PGID : %d \n",getpgid(getpid()));
 
 	if (bgChildPid == 0)
 	{		
-		sleep(20);
+		//sleep(5);
 		execvp(tokens[0],tokens);
 		printf("Shell : Incorrect command\n");
 		exit(1);
 	}
 	else
 	{
+		bgPid = bgChildPid;
         signal(SIGCHLD,childSignalHandler);
         processes[noOfProcesses++] = bgChildPid;	
 	}
@@ -141,11 +175,19 @@ void executeCommandInBackground(char **tokens)
 void ctrl_C_Handler(int sig)
 {
     int status;
-	if(kill(fgPid, 9)<0)
-    {
-        perror("Ctrl+C ");
-    }
-	printf(" Foreground process closed.\n");
+	if(fgPid > 0)
+	{
+		if(kill(fgPid, 9)<0)
+		{
+			perror("Ctrl+C ");
+		}
+		printf(" Foreground process closed.\n");
+		printf("PID : %d is now reaped.\n", fgPid);
+	}
+	else
+	{
+		printf(" No Foreground process is executing.\n");
+	}
 }
 
 int main(int argc, char* argv[]) {
@@ -154,8 +196,25 @@ int main(int argc, char* argv[]) {
 	int i,status;
 
     signal(SIGINT, ctrl_C_Handler);
+
 	while(1) 
 	{		
+
+		//TO GET THE "BACKGROUND PROCESS COMPLETED" AFTER PRESSING ENTER
+
+		// pid_t w = waitpid(bgPid, &status, WNOHANG);
+		// if(w == bgPid)
+		// {
+		// 	printf("Shell: Background process finished\n");
+		// 	for (int i = 0; i < noOfProcesses; i++)
+		// 	{
+		// 		if (processes[i] == bgPid)
+		// 		{
+		// 			processes[i] = -10;
+		// 		}			
+		// 	}
+		// }
+
 		/* BEGIN: TAKING INPUT */
 		bzero(line, sizeof(line));
 		printf("$ ");
@@ -180,12 +239,29 @@ int main(int argc, char* argv[]) {
 			printf("found token %s (remove this debug output later)\n", tokens[i]);
 		}
         */
-	   //int size = sizeof(tokens)/sizeof(tokens[0]);
+
 	   if(strcmp(tokens[noOfTokens-1],"&") != 0)
 	   {		
 			if (strcmp(tokens[0],"cd")==0)
 			{
-				chdir(tokens[1]);
+				// To check the syntax for cd command
+				if(!tokens[2])
+				{
+					if(isValidDirectory(tokens[1]))
+					{
+						chdir(tokens[1]);
+					}
+					else
+					{
+						printf("Please enter valid Directory name.\n");
+						continue;
+					}
+				}
+				else
+				{
+					printf("Shell: Incorrect command\n");
+					continue;
+				}
 			}
 			else if (strcmp(tokens[0],"exit") == 0)
 			{
@@ -210,5 +286,6 @@ int main(int argc, char* argv[]) {
 		free(tokens);
 
 	}
+
 	return 0;
 }
